+++ src/vm/pasopia/joypac2.h
@@ -19,7 +19,7 @@
 class JOYPAC2 : public PAC2DEV
 {
 private:
-	uint8* joy;
+	uint32* joy;
 public:
 	JOYPAC2(VM* parent_vm, EMU* parent_emu) : PAC2DEV(parent_vm, parent_emu) {}
 	~JOYPAC2() {}
